Revert "Improve app class without abstraction (#3036)"
This reverts commit 9b05a4ee16c8e628c31b854c20257204d815c3b6.
