+++ backend/common/src/main/java/ai/verta/modeldb/common/CommonUtils.java
@@ -17,9 +17,6 @@ import com.google.rpc.Status;
 import io.grpc.StatusRuntimeException;
 import io.grpc.protobuf.StatusProto;
 import io.grpc.stub.StreamObserver;
-import java.io.BufferedReader;
-import java.io.File;
-import java.io.FileReader;
 import java.net.SocketException;
 import java.util.ArrayList;
 import java.util.List;
@@ -297,33 +294,5 @@ public class CommonUtils {
     var query = handle.createQuery(queryStr);
     queryContext.getBinds().forEach(b -> b.accept(query));
     return query;
-  }
-
-  public static void resolvePortCollisionIfExists(String path) throws Exception {
-    File pidFile = new File(path);
-    if (pidFile.exists()) {
-      try (BufferedReader reader = new BufferedReader(new FileReader(pidFile))) {
-        String pidString = reader.readLine();
-        var pid = Long.parseLong(pidString);
-        var process = ProcessHandle.of(pid);
-        if (process.isPresent()) {
-          var processHandle = process.get();
-          LOGGER.warn("Port is already used by backend PID: {}", pid);
-          boolean destroyed = processHandle.destroy();
-          LOGGER.warn("Process kill completed `{}` for PID: {}", destroyed, pid);
-          processHandle = processHandle.onExit().get();
-          LOGGER.warn("Process is alive after kill: `{}`", processHandle.isAlive());
-        }
-      }
-    }
-  }
-
-  public static void cleanUpPIDFile() {
-    var path = System.getProperty(CommonConstants.USER_DIR) + "/" + CommonConstants.BACKEND_PID;
-    File pidFile = new File(path);
-    if (pidFile.exists()) {
-      pidFile.deleteOnExit();
-      LOGGER.trace(CommonConstants.BACKEND_PID + " file is deleted: {}", pidFile.exists());
-    }
   }
 }
