+++ backend/common/src/main/java/ai/verta/modeldb/common/config/Config.java
@@ -2,7 +2,6 @@ package ai.verta.modeldb.common.config;
 
 import static io.opentelemetry.api.common.AttributeKey.stringKey;
 
-import ai.verta.modeldb.common.CommonConstants;
 import ai.verta.modeldb.common.CommonUtils;
 import ai.verta.modeldb.common.exceptions.InternalErrorException;
 import ai.verta.modeldb.common.exceptions.ModelDBException;
@@ -239,11 +238,5 @@ public abstract class Config {
       openTelemetry = initializeOpenTelemetry();
     }
     return openTelemetry;
-  }
-
-  public boolean isMigration() {
-    return Boolean.parseBoolean(
-        Optional.ofNullable(System.getenv(CommonConstants.ENABLE_LIQUIBASE_MIGRATION_ENV_VAR))
-            .orElse("false"));
   }
 }
