--- backend/common/src/main/java/ai/verta/modeldb/common/configuration/GracefulShutdownBeans.java
@@ -1,27 +1,0 @@
-package ai.verta.modeldb.common.configuration;
-
-import ai.verta.modeldb.common.GracefulShutdown;
-import ai.verta.modeldb.common.config.Config;
-import org.springframework.boot.web.embedded.tomcat.TomcatServletWebServerFactory;
-import org.springframework.boot.web.servlet.server.ServletWebServerFactory;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class GracefulShutdownBeans {
-
-  @Bean
-  public GracefulShutdown gracefulShutdown(Config config) {
-    if (config.getSpringServer().getShutdownTimeout() == null) {
-      return new GracefulShutdown(30L);
-    }
-    return new GracefulShutdown(config.getSpringServer().getShutdownTimeout());
-  }
-
-  @Bean
-  public ServletWebServerFactory servletContainer(final GracefulShutdown gracefulShutdown) {
-    var factory = new TomcatServletWebServerFactory();
-    factory.addConnectorCustomizers(gracefulShutdown);
-    return factory;
-  }
-}
