+++ backend/src/main/java/ai/verta/modeldb/artifactStore/storageservice/s3/RefCountedS3Client.java
@@ -1,6 +1,6 @@
 package ai.verta.modeldb.artifactStore.storageservice.s3;
 
-import ai.verta.modeldb.configuration.CronJobUtils;
+import ai.verta.modeldb.cron_jobs.CronJobUtils;
 import com.amazonaws.auth.AWSCredentials;
 import com.amazonaws.services.s3.AmazonS3;
 import java.util.concurrent.atomic.AtomicInteger;
