+++ backend/src/main/java/ai/verta/modeldb/cron_jobs/CronJobUtils.java
@@ -1,28 +1,22 @@
-package ai.verta.modeldb.configuration;
+package ai.verta.modeldb.cron_jobs;
 
 import ai.verta.modeldb.ModelDBConstants;
 import ai.verta.modeldb.ServiceSet;
 import ai.verta.modeldb.artifactStore.ArtifactStoreDAODisabled;
 import ai.verta.modeldb.common.config.CronJobConfig;
 import ai.verta.modeldb.config.MDBConfig;
-import ai.verta.modeldb.cron_jobs.CleanUpEntitiesCron;
-import ai.verta.modeldb.cron_jobs.DeleteEntitiesCron;
-import ai.verta.modeldb.cron_jobs.PopulateEnvironmentInRunCron;
 import ai.verta.modeldb.utils.ModelDBUtils;
 import java.util.Map;
 import java.util.TimerTask;
 import java.util.concurrent.TimeUnit;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
 
-@Configuration
 public class CronJobUtils {
   private static final Logger LOGGER = LogManager.getLogger(CronJobUtils.class);
 
-  @Bean
-  public CronJobUtils initializeCronJobs(MDBConfig mdbConfig, ServiceSet services) {
+  public static void initializeCronJobs(MDBConfig mdbConfig, ServiceSet services) {
+
     LOGGER.info("Enter in CronJobUtils: initializeBasedOnConfig()");
     if (mdbConfig.getCron_job() != null) {
       for (Map.Entry<String, CronJobConfig> cronJob : mdbConfig.getCron_job().entrySet()) {
@@ -61,6 +55,5 @@ public class CronJobUtils {
       }
     }
     LOGGER.info("Exit from CronJobUtils: initializeBasedOnConfig()");
-    return this;
   }
 }
