+++ backend/src/main/java/ai/verta/modeldb/telemetry/TelemetryUtils.java
@@ -19,7 +19,8 @@ import org.apache.logging.log4j.Logger;
 
 public class TelemetryUtils {
   private static final Logger LOGGER = LogManager.getLogger(TelemetryUtils.class);
-  private static ModelDBHibernateUtil modelDBHibernateUtil;
+  private static final ModelDBHibernateUtil modelDBHibernateUtil =
+      ModelDBHibernateUtil.getInstance();
   private static boolean telemetryInitialized = false;
   public static String telemetryUniqueIdentifier = null;
   private String consumer = ModelDBConstants.TELEMETRY_CONSUMER_URL;
@@ -28,7 +29,6 @@ public class TelemetryUtils {
     if (consumer != null && !consumer.isEmpty()) {
       this.consumer = consumer;
     }
-    modelDBHibernateUtil = ModelDBHibernateUtil.getInstance();
     initializeTelemetry();
   }
