+++ backend/src/main/java/ai/verta/modeldb/utils/CommonHibernateUtil.java
@@ -44,7 +44,7 @@ public abstract class CommonHibernateUtil extends CommonDBUtil {
   // private HibernateStatisticsCollector hibernateStatisticsCollector;
 
   public Connection getConnection() throws SQLException {
-    return getSessionFactory()
+    return sessionFactory
         .getSessionFactoryOptions()
         .getServiceRegistry()
         .getService(ConnectionProvider.class)
